Fixing broken unit tests.
